% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramlegacy.R
\name{CloseRAM}
\alias{CloseRAM}
\title{A function to close connection to the RAM legacy DB}
\usage{
CloseRAM(con.drv = NULL)
}
\arguments{
\item{con.drv}{a list containing a non-expired connection to and driver for RAM legacy DB}
}
\description{
A function to close connection to the RAM legacy DB
}
\details{
closes databse connection using DBI-compliant connection via RPostgreSQL package
}
\examples{
\dontrun{
## make a connection
c <- ConnectRAM() #running without argumetns requires RAM.auth list set as global option
CloseRAM(c)
}
}

