% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramlegacy.R
\name{ConnectRAM}
\alias{ConnectRAM}
\title{A function to connect to the RAM legacy stock assessment database}
\usage{
ConnectRAM(auth.list = getOption("RAM.auth",
  stop("Need to provide a list of authorization args with elements host, user, password, dbname, port")),
  drv.name = "PostgreSQL")
}
\arguments{
\item{auth.list}{a list including authorization arguments including host, user, password, dbname, port (provided at link above)}

\item{drv.name}{database driver to use for DBI ('PostgreSQL' on mac/linux)}
}
\description{
A function to connect to the RAM legacy stock assessment database
}
\details{
constructs a databse connection using DBI-compliant connection via RPostgreSQL package
returns list of connection and driver.
(latest connection info:
http://ramlegacy.marinebiodiversity.ca/ram-legacy-stock-assessment-database/accessing-the-live-database)
before using be sure to read the fair use policy:
http://ramlegacy.marinebiodiversity.ca/ram-legacy-stock-assessment-database/ram-legacy-stock-assessment-database-fair-use-policy
}
\examples{
\dontrun{
# Can add lines to your .Rprofile file to store connection info provided at http://ramlegacy.marinebiodiversity.ca/ram-egacy-stock-assessment-database/accessing-the-live-database
options(RAM.auth =  alist(host="nautilus-vm.mathstat.dal.ca", user="srdbuser", password="srd6us3r!", dbname="srdb", port="5432"))
## make a connection
c <- ConnectRAM()
CloseRAM(c)
}
}

