% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramlegacy.R
\name{QueryRAM}
\alias{QueryRAM}
\title{A function to send SQL queries to the RAM legacy stock assessment database}
\usage{
QueryRAM(query, con.drv = NULL)
}
\arguments{
\item{query}{a valid SQL query executable against RAM Legacy DB}

\item{con.drv}{a list containing a non-expired connection to and driver for RAM legacy DB}
}
\description{
A function to send SQL queries to the RAM legacy stock assessment database
}
\details{
Runs SQL query against srdb (RAM Legacy DB) and returns result
before using be sure to read the fair use policy:
http://ramlegacy.marinebiodiversity.ca/ram-legacy-stock-assessment-database/ram-legacy-stock-assessment-database-fair-use-policy
}
\examples{
\dontrun{
## make a connection
c <- ConnectRAM()  #running without argumetns requires RAM.auth list set as global option
lme.df<-QueryRAM("select * from srdb.lmerefs", c)
CloseRAM(c)
}
}

